LPS-70677 Configurators are public API, we should export their package
