+++ modules/sdk/gradle-plugins-workspace/src/main/java/com/liferay/gradle/plugins/workspace/WorkspaceExtension.java
@@ -14,11 +14,11 @@
 
 package com.liferay.gradle.plugins.workspace;
 
-import com.liferay.gradle.plugins.workspace.internal.configurators.ModulesProjectConfigurator;
-import com.liferay.gradle.plugins.workspace.internal.configurators.PluginsProjectConfigurator;
-import com.liferay.gradle.plugins.workspace.internal.configurators.RootProjectConfigurator;
-import com.liferay.gradle.plugins.workspace.internal.configurators.ThemesProjectConfigurator;
-import com.liferay.gradle.plugins.workspace.internal.configurators.WarsProjectConfigurator;
+import com.liferay.gradle.plugins.workspace.configurators.ModulesProjectConfigurator;
+import com.liferay.gradle.plugins.workspace.configurators.PluginsProjectConfigurator;
+import com.liferay.gradle.plugins.workspace.configurators.RootProjectConfigurator;
+import com.liferay.gradle.plugins.workspace.configurators.ThemesProjectConfigurator;
+import com.liferay.gradle.plugins.workspace.configurators.WarsProjectConfigurator;
 import com.liferay.gradle.plugins.workspace.internal.util.GradleUtil;
 
 import groovy.lang.MissingPropertyException;
